git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2376 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Graphics/GRFrame.h
@@ -35,7 +35,7 @@ public:
 
 	virtual GRFrameIf* GetParent(){ return parent->Cast(); }
 	virtual void SetParent(GRFrameIf* fr);
-	virtual int NChildren(){ return children.size(); }
+	virtual int NChildren(){ return (int)children.size(); }
 	virtual GRVisualIf** GetChildren(){ return (GRVisualIf**)&*children.begin(); }
 	virtual void SetNameManager(NameManager* n);
 	virtual GRSceneIf* GetScene();
